fix(ort_types): condensing doc comments and fix typo

+++ ort_types/src/imu.rs
@@ -1,7 +1,7 @@
-//! IMU ample types, shared between the firmware and the desktop app.
+//! IMU sample types, shared between the firmware and the desktop app.
 //!
 //! These are raw register units as they come off the MPU6050. Converting them
-//! to physical units is the firmware's job — see `firmware/src/imu_math.rs`.
+//! to physical units is the firmware's job (see `firmware/src/imu_math.rs`).
 
 /// Accel, temperature, and gyro registers are contiguous starting at
 /// `ACCEL_OUT_H`, so one burst read picks up all 7 samples.
@@ -10,7 +10,6 @@ pub const READ_BUF_SIZE: usize = 14;
 /// One full sample set, in raw register units.
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "specta", derive(specta::Type))]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ImuData {
     pub accel_x: i16,
@@ -40,8 +39,13 @@ impl ImuData {
     }
 }
 
+/// Combines a high/low register pair into one signed sample.
+///
+/// The widening to `u16` has to happen before the shift: `u8 << 8` discards
+/// every bit and the compiler rejects it. `as i16` then reinterprets the
+/// two's-complement bit pattern the MPU6050 produces.
 fn combine_bytes(high: u8, low: u8) -> i16 {
-    (high << 8 | low) as i16
+    (((high as u16) << 8) | low as u16) as i16
 }
 
 #[cfg(test)]
